+++ src/main/kotlin/com/github/warningimhack3r/npmupdatedependencies/backend/models/Deprecation.kt
@@ -1,9 +1,22 @@
 package com.github.warningimhack3r.npmupdatedependencies.backend.models
 
 data class Deprecation(
+    val kind: Kind,
     val reason: String,
     val replacement: Replacement?
 ) {
+    enum class Kind {
+        DEPRECATED,
+        UNMAINTAINED;
+
+        override fun toString(): String {
+            return when (this) {
+                DEPRECATED -> "Deprecated"
+                UNMAINTAINED -> "Unmaintained"
+            }
+        }
+    }
+
     data class Replacement(
         val name: String,
         val version: String
@@ -11,12 +24,14 @@ data class Deprecation(
 
     enum class Action {
         REPLACE,
-        REMOVE;
+        REMOVE,
+        IGNORE;
 
         override fun toString(): String {
             return when (this) {
                 REPLACE -> "Replace"
                 REMOVE -> "Remove"
+                IGNORE -> "Ignore"
             }
         }