+++ src/main/kotlin/com/github/warningimhack3r/npmupdatedependencies/settings/NUDSettingsComponent.kt
@@ -1,20 +1,15 @@
 package com.github.warningimhack3r.npmupdatedependencies.settings
 
-import com.github.warningimhack3r.npmupdatedependencies.backend.engine.NUDState
 import com.github.warningimhack3r.npmupdatedependencies.backend.models.Deprecation
 import com.github.warningimhack3r.npmupdatedependencies.backend.models.Versions
 import com.github.warningimhack3r.npmupdatedependencies.ui.statusbar.StatusBarMode
-import com.intellij.codeInsight.daemon.DaemonCodeAnalyzer
 import com.intellij.ide.DataManager
 import com.intellij.openapi.application.ApplicationNamesInfo
 import com.intellij.openapi.application.ModalityState
 import com.intellij.openapi.application.runInEdt
-import com.intellij.openapi.fileEditor.FileEditorManager
 import com.intellij.openapi.options.ex.Settings
-import com.intellij.openapi.project.ProjectManager
 import com.intellij.openapi.ui.DialogBuilder
 import com.intellij.openapi.ui.DialogWrapper
-import com.intellij.psi.PsiDocumentManager
 import com.intellij.ui.ToolbarDecorator
 import com.intellij.ui.components.JBCheckBox
 import com.intellij.ui.dsl.builder.*
@@ -133,20 +128,38 @@ class NUDSettingsComponent {
                     .bindItem(settings::defaultUpdateType.toMutableProperty())
             }
             row("Default deprecation action:") {
-                comboBox(Deprecation.Action.entries.toList())
+                comboBox(Deprecation.Action.entries.toList().filter { it != Deprecation.Action.IGNORE })
                     .bindItem(settings::defaultDeprecationAction.toMutableProperty())
             }
         }
         group("Deprecations") {
+            lateinit var bannerEnabled: Cell<JBCheckBox>
             row {
-                checkBox("Show deprecation banner")
+                bannerEnabled = checkBox("Show deprecation banner")
                     .comment("Show a warning banner when at least one dependency is deprecated.")
                     .bindSelected(settings::showDeprecationBanner)
+            }
+            indent {
+                row {
+                    checkBox("Include \"unmaintained\" packages")
+                        .comment("Also show the banner when a package hasn't been updated in a specified amount of time.")
+                        .bindSelected(settings::bannerIncludesUnmaintained)
+                }.enabledIf(bannerEnabled.selected)
             }
             row {
                 checkBox("Automatically reorder dependencies")
                     .comment("Reorder dependencies after replacing deprecated ones.<br>Useful when a new dependency starts with a different letter than the old one.")
                     .bindSelected(settings::autoReorderDependencies)
+            }
+            row("Days until a package is considered unmaintained:") {
+                spinner(0..365 * 10)
+                    .comment(
+                        """
+                        Control how many days a package can go without an update before it's considered "likely unmaintained". Set to 0 to disable.<br>
+                        <em>A package might still be maintained even if it hasn't been updated in a while. You can adjust this value to your liking, or even exclude packages from this check if you know they're still maintained.</em>
+                        """.trimIndent()
+                    )
+                    .bindIntValue(settings::unmaintainedDays)
             }
         }
         group("Parallelism") {
@@ -204,22 +217,13 @@ class NUDSettingsComponent {
                                 values.isNotEmpty()
                             }
                         }.toMutableMap()
-                        ProjectManager.getInstance().openProjects.forEach { project ->
-                            // Clear the cache for packages with excluded versions
-                            settings.excludedVersions.keys.forEach { packageName ->
-                                NUDState.getInstance(project).availableUpdates.remove(packageName)
-                            }
-                            // if project's currently open file is package.json, re-analyze it
-                            FileEditorManager.getInstance(project).selectedTextEditor?.let { editor ->
-                                PsiDocumentManager.getInstance(project).getPsiFile(editor.document)?.let { file ->
-                                    if (file.name == "package.json") {
-                                        DaemonCodeAnalyzer.getInstance(project).restart(file)
-                                    }
-                                }
-                            }
-                        }
                     }.showAndGet()
                 }
+            }
+            row("Packages excluded from the \"unmaintained\" check:") {
+                textField()
+                    .comment("Comma-separated list of package names that should be excluded from the \"unmaintained\" check.")
+                    .bindText(settings::excludedUnmaintainedPackages)
             }
         }