+++ src/main/kotlin/com/github/warningimhack3r/npmupdatedependencies/ui/helpers/NUDHelper.kt
@@ -1,9 +1,12 @@
 package com.github.warningimhack3r.npmupdatedependencies.ui.helpers
 
+import com.intellij.codeInsight.daemon.DaemonCodeAnalyzer
 import com.intellij.lang.Language
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.command.WriteCommandAction
+import com.intellij.openapi.fileEditor.FileEditorManager
 import com.intellij.openapi.project.Project
+import com.intellij.psi.PsiDocumentManager
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiFileFactory
@@ -54,5 +57,15 @@ object NUDHelper {
             sibling = sibling.prevSibling
         }
         return null
+    }
+
+    fun reanalyzePackageJsonIfOpen(project: Project) {
+        FileEditorManager.getInstance(project).selectedTextEditor?.let { editor ->
+            PsiDocumentManager.getInstance(project).getPsiFile(editor.document)?.let { file ->
+                if (file.name == "package.json") {
+                    DaemonCodeAnalyzer.getInstance(project).restart(file)
+                }
+            }
+        }
     }
 }